refactor(zestymarkdownparser.js): added components
