+++ src/components/markdown-styling/components/MDH3.js
@@ -1,0 +1,18 @@
+import { Typography, Box } from '@mui/material';
+
+export const MDH3 = ({ node }) => {
+  return (
+    <Box>
+      <Typography
+        variant="h6"
+        component={'h3'}
+        id={node.children[0].value
+          ?.replace(/[^\w\s]/gi, '')
+          ?.replace(/\s+/g, '-')
+          ?.toLowerCase()}
+      >
+        {node.children[0].value}
+      </Typography>
+    </Box>
+  );
+};
