+++ src/components/markdown-styling/components/index.js
@@ -1,0 +1,6 @@
+export * from './MDH1';
+export * from './MDH2';
+export * from './MDH3';
+export * from './MDBlockquote';
+export * from './MDParagraph';
+export * from './MDImage';
